requirements.txt
blinker==1.9.0
branca==0.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
folium==0.20.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.0
python-dotenv==1.1.0
requests==2.32.4
urllib3==2.5.0
Werkzeug==3.1.3
xyzservices==2025.4.0
